Add support for Kibana 4
This changeset adds support for Kibana 4. Kibana 4 is very different
from Kibana 3.x in that it is a JVM based application. Previous versions
of Kibana were standalone JavaScript applications that only interacted
with ElasticSearch as a backend.

In addition, the example in this role was updated to use a standalone
ElasticSearch instance vs. the embedded ElasticSearch in Logstash.